Paper Slide-In Optimization
The slide in animation was appearing at the bottom left of the entire site, I messed with the js and css in order to get it to slide onto the recipe box image. I couldn't get it to slide from the lid, but if we were to maybe break that recipe box down into layers I could find some way to stack them?
-Brooke